Fix typo and remove duplicate table rows
I'm assuming "tun" was supposed to be "run" though the mistake was made twice which made me question if it's an acronym I'm unfamiliar with

The last two table rows duplicated the first two. It looks like there are other examples in https://github.com/prisma/prisma-examples/tree/master/typescript that maybe could be added? But I'll let y'all figure out which to add and what description to use, etc.