Updates schema to match example project.
Pasting in the code in the **Rename the relation fields for easy access** currently breaks the tutorial. The schema in the example project appears to have changed since this article was originally written. This PR updates the code to its current schema and applies the changes described in **Rename the relation fields for easy access** to this new schema.