Remove outdated note
The note states "Note: generate is called under the hood by default, after running prisma migrate dev. If the prisma-client-js generator is defined in your schema, this will check if @prisma/client is installed and install it if it's missing." however when I ran prisma migrate dev, @prisma/client did not get installed, furthermore installing @prisma/client is actually the next step in the tutorial so this note seems outdated?