Update 150-db-push.mdx
Added a Note to tell users of PlanetScale that we recommend `db push` instead of `migrate`. Added a link to the Getting Started docs... but after the PlanetScale-specific docs are published, we can change that link to be more specific. And we can add more details to the upcoming Guide as well, about WHY we recommend `db push`. See [Issue #7282](https://github.com/prisma/prisma/issues/7292#issuecomment-988624732).