Add `--transpile-only` flag to seed script
Failing to use the `--transpile-only` flag in this context is a catastrophic mistake that consumes massive amounts of RAM when typechecking the seed script(s) unnecessarily, in addition to transpiling them, thereby completely crashing Node in any container with limited RAM i.e. Docker, Heroku, etc. when trying to seed.