Add indexes to the explicit relations example
It's a standard optimization to index the foreign keys on the explicit relations table, so I think it would be a good idea to improve the example by adding the indexes here. This is the only place I could find in the docs of an explicit relations example, so we should help newcomers see a perfect implementation of how to do it! What do you think?