Improve excluding fields example to not mutate data (#4794)
The previous implementation mutated the passed object. This updated version does not. The issue here is that in this example `password` is deleted from the passed object, `user`, causing a call to password to fail after exclude is called. E.G.: `user.password.trim()` would fail with "cannot call 'trim' on 'undefined'".

Co-authored-by: Alex Ruheni <33921841+ruheni@users.noreply.github.com>