fix(prisma-client-transactions-guide): Make $transaction([]) explicit
Right now the document refers to the batch transaction API only as "`$transaction` API". But interactive transactions of course also use `$transaction()`, just with a different parameter (functions instead of array).

Calling it the "`$transaction([])` API" at least lets us differentiate.

This needs to be cleaned up on a higher level though as well.