feat: new mdx component for version info
Version info in full sentence or paragraphs that has version related information. Right now that is all over the place, sometimes separate or sometimes in half sentences. This new component would be a way to force that it is separate and the rest of the text also works without the version information.