Refactor Docs Page: Remove Unnecessary Code Snippet Update 07-mongodb.mdx
This pull request addresses an issue in the documentation page where an unnecessary code snippet was present. The code snippet has been identified and removed to enhance the clarity and conciseness of the documentation.  there is a one example of user schema in mongodb section that is written two times