Add documentation page for MariaDB
The existing page about MySQL does not mention MariaDB at all, and thus it
is unclear to users of MariaDB if Prisma properly supports MariaDB or not.
With a new MariaDB-specific page it is easy for users to find that Prisma
does indeed support MariaDB, and even has MariaDB in the CI for Prisma.

Creating two separate pages for MySQL and MariaDB is useful as future
versions of MySQL 8.4 and MariaDB 11.5 are likely to start diverging and
thus Prisma users need to have separate pages that give them instructions
that are guaranteed to work specifically with the database they use.